casting as float to make qjit compatible

+++ pennylane/templates/subroutines/amplitude_amplification.py
@@ -36,10 +36,10 @@ def _get_fixed_point_angles(iters, p_min):
     gamma = np.cos(np.arccos(1 / delta, dtype=np.complex128) / iters, dtype=np.complex128) ** -1
 
     alphas = [
-        float(2 * np.arctan(1 / (np.tan(2 * np.pi * j / iters) * np.sqrt(1 - gamma**2))))
+        2 * np.arctan(1 / (np.tan(2 * np.pi * j / iters) * np.sqrt(1 - gamma**2)))
         for j in range(1, iters // 2 + 1)
     ]
-    betas = [float(-alphas[-j]) for j in range(1, iters // 2 + 1)]
+    betas = [-alphas[-j] for j in range(1, iters // 2 + 1)]
     return alphas[: iters // 2], betas[: iters // 2]
